Function get_approval_comment_body update for dust
This is to indicate the user is being rewarded via beneficiary rewards vs an upvote when their flag does not meet our SBD minimum payout threshold. Also, enclosed the code pulling the category in a try - except so that key errors are no longer a bot "show-stopper" (requiring admin intervention to restart).