Queue toggle and Unqueue functions
These should not be needed when queue is functioning correctly and are in place as contingencies when a server admin is not available to troubleshoot. Queue toggle bypasses the queuing mechanism allowing approving flags for upvotes irrespective of bot voting mana. Unqueue may be used if a comment is noticed to have already been voted on but, for whatever reason, the SQL did not properly execute. This has not really been an issue since the pre-vote check has been implemented.