Redesigned Flaggable Post Query
It now displays all flagged content that is less than 5 days old. This helps us to focus our voting power where it is most effective and avoid the late voting penalty. The entire query for flaggable posts is below. If you have any ideas, feel free to drop a comment.