Batch Follow On Approval
This function allows mods to pass a link to the @steemflagrewards bot approval comment and submit all child comments for approval. This greatly improved the amount of time it takes when a mod finds a post that has many flag mentions. It's a great timesaver so we can focus more on abuse fighting and related development. Note: The insert_mention function will be in the next commit.